feat: Created initial routes for handling source connector behaviour
route /verify for testing db connection and return an object container schemas/tables/columns info
route /source  which handles connector setup to kafka connect

Also created a general error handler in the index.ts file and a general http error class. Some refactoring and polishing still to be done.

Arising issues: when initianing a connector, if the db has an existing replication slot name, it will create the connector istance and reply with a 200 code but fails gracefully inside kafka connect.

Need to reach a consensus on how to define a topic prefix as well.