fix(debezium): :bug: Fix Debezium source connector for converting numeric/decimal values.
By default, Debezium converts numeric/decimal Postgres values to either BYTES or a STRUCT that has info on the value's scale & the value in BYTES. This is difficult to manage & would require some complex conversion to handle appropriately. Instead, we can just tell the Debezium connector to convert numeric/decimal values to float64 with the `decimal.handling.mode` set to `double`.