feat(redis): :sparkles: Ensure `processKafkaMessage` properly handles tables without primary keys.
Added a custom error `NoPrimaryKeyError` within `utils.ts` - this is for when we encounter a table with no primary keys, which we do not want to allow. Throw this error in the `processKafkaMessage` method within `redis.ts` if we do encounter these tables. We will most definitely want to prevent tables without primary keys earlier in the pipeline, but include this error handling here for now.

Also made the logic to detect `TRUNCATE` operates more specific i.e., `if (key === null && operation === TRUNCATE)`.