refactor(types): :label: Remove type assertions/casts from routes.
Created a generic type `TypedRequest` that can be used for both source and sink requests to specify what properties are in their bodies.
We may want to revisit later and make all properties of the `SourceRequestBody` and `SinkRequestBody` to be optional and then do some validation to confirm the expected properties are actually in the response body.