fix(ui): :bug: Hide tables with no primary keys in SELECT DATA form and set their `selected` property to `false`.
Since we must send excluded tables to the backend for the Debezium configuration, we must include them in the `formData` state. I've added a `visible` property to hide tables with no primary keys and set their `selected` property to `false` so they will be sent to the backend in the list of tables to exclude.