refactor(ui): :lipstick: Handle validation: users expected now to select at least one table. Also Primary keys columns switches are disabled by default (must always be selected)
table selection of at least one item was needed in order to handle the backed logic appropriately given the latest changes in the backed for debezium configuration. Not allowing to disable the primary key likewise was needed to handle the discovered bug for the initial debezium snapshot process

Split the ListItemWithSwitch into two different components as now we're dealing with different logic in each of those components. At the same time, deleted the unneeded previous component as well. Updating typing as well.

Co-authored-by: Cameron Bibb <cameron.bibb@icloud.com>