fix(redis): :bug: Allow Redis client to reconnect after a connection timeout occurs.
Often by default, Redis servers have a connection timeout of ~5 minutes. Once the connection times out, a consumer's `Redis` instance is disconnected. Previously, we were throwing a `RedisError` that was not handled, causing this timeout to crash our app. Instead, by logging the error, the Redis client is allowed to reconnect after a connection timeout and not crash our app.

Co-authored-by: Arno Cai <acai367@gmail.com>